requirements.txt
Django>=4.2
djangorestframework>=3.14
pymongo>=4.6
firebase-admin>=6.4
python-dotenv>=1.0
django-cors-headers>=4.3
